+++ js/main.js
@@ -1,2 +1,3 @@
-import {similarAdvertisement} from './main.js';
+import {SIMILAR_ADVERTISEMENT_COUNT, createAdvertisement} from './data.js';
+const similarAdvertisement = new Array(SIMILAR_ADVERTISEMENT_COUNT).fill(null).map(() => createAdvertisement());
 similarAdvertisement;
